Update lots of code to standard services from zFunctions
