+++ src/services/dateServices.ts
@@ -1,6 +1,6 @@
 
 
-import { IUser} from '../webparts/pivottiles7/components/IReUsableInterfaces';
+import { IUser} from './IReUsableInterfaces';
 
 export const msPerMin = 60000;
 export const msPerHr = 3600000;
