+++ src/services/stringServices.ts
@@ -1,5 +1,25 @@
+/***
+ *    .d888b.  .d88b.  .d888b.  .d88b.          db .d888b.         db   j88D                     
+ *    VP  `8D .8P  88. VP  `8D .8P  88.        o88 VP  `8D        o88  j8~88                     
+ *       odD' 88  d'88    odD' 88  d'88         88    odD'         88 j8' 88                     
+ *     .88'   88 d' 88  .88'   88 d' 88 C8888D  88  .88'   C8888D  88 V88888D                    
+ *    j88.    `88  d8' j88.    `88  d8'         88 j88.            88     88                     
+ *    888888D  `Y88P'  888888D  `Y88P'          VP 888888D         VP     VP                     
+ *                                                                                               
+ *                                                                                               
+ *    d8888b. d888888b db    db  .d88b.  d888888b      d888888b d888888b db      d88888b .d8888. 
+ *    88  `8D   `88'   88    88 .8P  Y8. `~~88~~'      `~~88~~'   `88'   88      88'     88'  YP 
+ *    88oodD'    88    Y8    8P 88    88    88            88       88    88      88ooooo `8bo.   
+ *    88~~~      88    `8b  d8' 88    88    88            88       88    88      88~~~~~   `Y8b. 
+ *    88        .88.    `8bd8'  `8b  d8'    88            88      .88.   88booo. 88.     db   8D 
+ *    88      Y888888P    YP     `Y88P'     YP            YP    Y888888P Y88888P Y88888P `8888Y' 
+ *                                                                                               
+ *                                                                                               
+ */
+
 import * as React from 'react';
 
+import { sortStringArray , sortNumberArray } from './arrayServices';
 //https://stackoverflow.com/a/2970667/4210807
 export function camelize(str,firstCap: boolean) {
 
@@ -132,4 +152,50 @@ export function encodeDecodeString( str : string , doThis: 'encode' | 'decode') 
 
   return newStr;
 
+}
+
+/** was: cleanProjEditOptions from TrackMyTime7
+ * This function takes a string with ;, converts to array of strings and removes empty elements (like if ; is at the end.)
+ * 
+ * example input:   ";test;this;string;;now;"
+ * example result:  ['test','this','string','now']
+ * 
+ * @param str
+ */
+export function getStringArrayFromString ( str : string, delim: string, removeEmpty: boolean, sort: 'asc' | 'dec' | null ) : string[] {
+
+    if (str == null ) { return null; }
+    else if (  delim == null || delim == '' ) { return [ str ]; }
+  
+    let arr : string[] = str.split( delim );
+
+    arr = sortStringArray( arr, sort );
+
+    let finalStringArray : string[] = [];
+
+    if ( removeEmpty === true ) {
+        finalStringArray = arr.filter( (el) => {
+            return el != null;
+        });
+    } else {
+        finalStringArray = arr;
+    }
+
+    return finalStringArray;
+  
+}
+
+/** was originally copied from cleanProjEditOptions from TrackMyTime7
+ * This function takes a string with ;, converts to array of strings and removes empty elements (like if ; is at the end.)
+ * 
+ * example input:   ";test;this;string;;now;"
+ * example result:  "test;this;string;now"
+ * @param str
+ */
+
+export function cleanEmptyElementsFromString ( str : string, delim: string, removeEmpty: boolean, sort: 'asc' | 'dec' | null ) : string {
+
+    let stringArray : string[] = getStringArrayFromString( str, delim, removeEmpty, sort );
+    return stringArray.join(';');
+  
 }
